Code check

+++ frameworks/surface/src/egl_manager.cpp
@@ -20,31 +20,6 @@
 #include <unistd.h>
 
 #include "buffer_log.h"
-// #include "shared/platform.h"
-
-#define CHECK_NOT_INIT(ret)                 \
-    if (!initFlag_) {                       \
-        BLOGE("EglManager is not init.");   \
-        return ret;                         \
-    }
-
-#define CHECK_NOT_INIT_VOID()               \
-    if (!initFlag_) {                       \
-        BLOGE("EglManager is not init.");   \
-        return;                             \
-    }
-
-#define CHECK_PARAM_NULLPTR(param)          \
-    if (param == nullptr) {                 \
-        BLOGE("param is nullptr.");         \
-        return SURFACE_ERROR_INIT;          \
-    }
-
-#define CHECK_PARAM_NULLPTR_RET(param, ret) \
-    if (param == nullptr) {                 \
-        BLOGE("param is nullptr.");         \
-        return ret;                         \
-    }
 
 namespace OHOS {
 namespace {
@@ -70,7 +45,6 @@ constexpr const char *EGL_DUP_NATIVE_FENCE_FD_ANDROID = "eglDupNativeFenceFDANDR
 constexpr const char *EGL_KHR_WAIT_SYNC = "EGL_KHR_wait_sync";
 constexpr const char *EGL_WAIT_SYNC_KHR = "eglWaitSyncKHR";
 constexpr const char *GL_OES_EGL_IMAGE = "GL_OES_EGL_image";
-
 } // namespace
 
 EglManager::EglManager()
@@ -131,7 +105,6 @@ SurfaceError EglManager::GbmInit()
     }
 
     device_ = gbm_create_device(drmFd_);
-
     if (device_ == nullptr) {
         BLOGE("Failed to create gbm device.");
         return SURFACE_ERROR_INIT;
@@ -143,35 +116,34 @@ SurfaceError EglManager::GbmInit()
 namespace {
 static bool CheckEglExtension(const char *extensions, const char *extension)
 {
-	size_t extlen = strlen(extension);
-	const char *end = extensions + strlen(extensions);
+    size_t extlen = strlen(extension);
+    const char *end = extensions + strlen(extensions);
 
-	while (extensions < end) {
-		size_t n = 0;
+    while (extensions < end) {
+        size_t n = 0;
 
-		/* Skip whitespaces, if any */
-		if (*extensions == CHARACTER_WHITESPACE) {
-			extensions++;
-			continue;
-		}
+        /* Skip whitespaces, if any */
+        if (*extensions == CHARACTER_WHITESPACE) {
+            extensions++;
+            continue;
+        }
 
-		n = strcspn(extensions, CHARACTER_STRING_WHITESPACE);
-
-		/* Compare strings */
-		if (n == extlen && strncmp(extension, extensions, n) == 0) {
+        n = strcspn(extensions, CHARACTER_STRING_WHITESPACE);
+        /* Compare strings */
+        if (n == extlen && strncmp(extension, extensions, n) == 0) {
             return true; /* Found */
         }
-		extensions += n;
-	}
-	/* Not found */
-	return false;
+        extensions += n;
+    }
+    /* Not found */
+    return false;
 }
 
-static EGLDisplay GetPlatformEglDisplay(EGLenum platform, void *native_display, const EGLint *attrib_list)
+static EGLDisplay GetPlatformEglDisplay(EGLenum platform, EGLDisplay native_display, const EGLint *attrib_list)
 {
-	static GetPlatformDisplayExt eglGetPlatformDisplayExt = NULL;
+    static GetPlatformDisplayExt eglGetPlatformDisplayExt = NULL;
 
-	if (!eglGetPlatformDisplayExt) {
+    if (!eglGetPlatformDisplayExt) {
         const char *extensions = eglQueryString(EGL_NO_DISPLAY, EGL_EXTENSIONS);
         if (extensions &&
             (CheckEglExtension(extensions, EGL_EXT_PLATFORM_WAYLAND) ||
@@ -179,19 +151,22 @@ static EGLDisplay GetPlatformEglDisplay(EGLenum platform, void *native_display, 
             eglGetPlatformDisplayExt =
                 (GetPlatformDisplayExt)eglGetProcAddress(EGL_GET_PLATFORM_DISPLAY_EXT);
         }
-	}
+    }
 
-	if (eglGetPlatformDisplayExt) {
+    if (eglGetPlatformDisplayExt) {
         return eglGetPlatformDisplayExt(platform, native_display, attrib_list);
     }
-	return eglGetDisplay((EGLNativeDisplayType) native_display);
+    return eglGetDisplay((EGLNativeDisplayType) native_display);
 }
 }
 
 SurfaceError EglManager::EglCheckExt()
 {
     const char *eglExtensions = eglQueryString(display_, EGL_EXTENSIONS);
-    CHECK_PARAM_NULLPTR(eglExtensions);
+    if (eglExtensions == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
 
     if (!CheckEglExtension(eglExtensions, EGL_EXT_IMAGE_DMA_BUF_IMPORT)) {
         BLOGE("EGL_EXT_image_dma_buf_import not supported");
@@ -229,34 +204,61 @@ SurfaceError EglManager::EglCheckExt()
 SurfaceError EglManager::EglFuncInit()
 {
     const char *eglExtensions = eglQueryString(display_, EGL_EXTENSIONS);
-    CHECK_PARAM_NULLPTR(eglExtensions);
+    if (eglExtensions == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
 
     createImage_ = (EglCreateImageFunc)eglGetProcAddress(EGL_CREATE_IMAGE_KHR);
-    CHECK_PARAM_NULLPTR(createImage_);
+    if (createImage_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
 
     destroyImage_ = (EglDestroyImageFunc)eglGetProcAddress(EGL_DESTROY_IMAGE_KHR);
-    CHECK_PARAM_NULLPTR(destroyImage_);
+    if (destroyImage_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
 
     imageTargetTexture2d_ = (EglImageTargetTexture2DFunc)eglGetProcAddress(EGL_IMAGE_TARGET_TEXTURE2DOES);
-    CHECK_PARAM_NULLPTR(imageTargetTexture2d_);
+    if (imageTargetTexture2d_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
     
     if (CheckEglExtension(eglExtensions, EGL_KHR_FENCE_SYNC)) {
         createSync_ = (EglCreateSyncFunc)eglGetProcAddress(EGL_CREATE_SYNC_KHR);
-        CHECK_PARAM_NULLPTR(createSync_);
+        if (createSync_ == nullptr) {
+            BLOGE("param is nullptr.");
+            return SURFACE_ERROR_INIT;
+        }
         
         destroySync_ = (EglDestroySyncFunc)eglGetProcAddress(EGL_DESTROY_SYNC_KHR);
-        CHECK_PARAM_NULLPTR(destroySync_);
+        if (destroySync_ == nullptr) {
+            BLOGE("param is nullptr.");
+            return SURFACE_ERROR_INIT;
+        }
 
         clientWaitSync_ = (EglClientWaitSyncFunc)eglGetProcAddress(EGL_CLIENT_WAIT_SYNC_KHR);
-        CHECK_PARAM_NULLPTR(clientWaitSync_);
+        if (clientWaitSync_ == nullptr) {
+            BLOGE("param is nullptr.");
+            return SURFACE_ERROR_INIT;
+        }
 
         dupNativeFenceFd_ = (EglDupNativeFenceFdFunc)eglGetProcAddress(EGL_DUP_NATIVE_FENCE_FD_ANDROID);
-        CHECK_PARAM_NULLPTR(dupNativeFenceFd_);
+        if (dupNativeFenceFd_ == nullptr) {
+            BLOGE("param is nullptr.");
+            return SURFACE_ERROR_INIT;
+        }
     }
 
     if (CheckEglExtension(eglExtensions, EGL_KHR_WAIT_SYNC)) {
         waitSync_ = (EglWaitSyncFunc)eglGetProcAddress(EGL_WAIT_SYNC_KHR);
-        CHECK_PARAM_NULLPTR(waitSync_);
+        if (waitSync_ == nullptr) {
+            BLOGE("param is nullptr.");
+            return SURFACE_ERROR_INIT;
+        }
     }
     return SURFACE_ERROR_OK;
 }
@@ -304,7 +306,10 @@ SurfaceError EglManager::EglInit(EGLContext ctx)
     eglMakeCurrent(display_, EGL_NO_SURFACE, EGL_NO_SURFACE, context_);
 
     const char *glExtensions = (const char *) glGetString(GL_EXTENSIONS);
-    CHECK_PARAM_NULLPTR(glExtensions);
+    if (glExtensions == nullptr) {
+        BLOGE("param is nullptr.");
+        return SURFACE_ERROR_INIT;
+    }
 
     if (!CheckEglExtension(glExtensions, GL_OES_EGL_IMAGE)) {
         BLOGE("GL_OES_EGL_image not supported");
@@ -342,86 +347,127 @@ void EglManager::Deinit()
 
 EGLDisplay EglManager::GetEGLDisplay() const
 {
-    CHECK_NOT_INIT(EGL_NO_DISPLAY);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_NO_DISPLAY;
+    }
     return display_;
 }
 
 EGLContext EglManager::GetEGLContext() const
 {
-    CHECK_NOT_INIT(EGL_NO_CONTEXT);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_NO_CONTEXT;
+    }
     return context_;
 }
 
 EGLImageKHR EglManager::EglCreateImage(EGLContext ctx, EGLenum target, EGLClientBuffer buffer, const EGLint *attribList)
 {
-    CHECK_NOT_INIT(EGL_NO_IMAGE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_NO_IMAGE;
+    }
     return createImage_(display_, ctx, target, buffer, attribList);
 }
 
 EGLImageKHR EglManager::EglCreateImage(EGLenum target, EGLClientBuffer buffer, const EGLint *attribList)
 {
-    CHECK_NOT_INIT(EGL_NO_IMAGE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_NO_IMAGE;
+    }
     return createImage_(display_, context_, target, buffer, attribList);
 }
 
 EGLBoolean EglManager::EglDestroyImage(EGLImageKHR image)
 {
-    CHECK_NOT_INIT(EGL_FALSE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_FALSE;
+    }
     return destroyImage_(display_, image);
 }
 
 void EglManager::EglImageTargetTexture2D(GLenum target, GLeglImageOES image)
 {
-    CHECK_NOT_INIT_VOID();
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return;
+    }
     return imageTargetTexture2d_(target, image);
 }
 
 EGLSyncKHR EglManager::EglCreateSync(EGLenum type, const EGLint *attribList)
 {
-    CHECK_PARAM_NULLPTR_RET(createSync_, EGL_NO_SYNC_KHR);
+    if (createSync_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return EGL_NO_SYNC_KHR;
+    }
     return createSync_(display_, type, attribList);
 }
 
 EGLint EglManager::EglWaitSync(EGLSyncKHR sync, EGLint flags)
 {
-    CHECK_PARAM_NULLPTR_RET(waitSync_, EGL_FALSE);
+    if (waitSync_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return EGL_FALSE;
+    }
     return waitSync_(display_, sync, flags);
 }
 
 EGLint EglManager::EglClientWaitSync(EGLSyncKHR sync, EGLint flags, EGLTimeKHR timeout)
 {
-    CHECK_PARAM_NULLPTR_RET(clientWaitSync_, EGL_FALSE);
+    if (clientWaitSync_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return EGL_FALSE;
+    }
     return clientWaitSync_(display_, sync, flags, timeout);
 }
 
 EGLBoolean EglManager::EglDestroySync(EGLSyncKHR sync)
 {
-    CHECK_PARAM_NULLPTR_RET(destroySync_, EGL_FALSE);
+    if (destroySync_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return EGL_FALSE;
+    }
     return destroySync_(display_, sync);
 }
 
 EGLint EglManager::EglDupNativeFenceFd(EGLSyncKHR sync)
 {
-    CHECK_PARAM_NULLPTR_RET(dupNativeFenceFd_, EGL_DONT_CARE);
+    if (dupNativeFenceFd_ == nullptr) {
+        BLOGE("param is nullptr.");
+        return EGL_DONT_CARE;
+    }
     return dupNativeFenceFd_(display_, sync);
 }
 
 EGLBoolean EglManager::EglMakeCurrent()
 {
-    CHECK_NOT_INIT(EGL_FALSE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_FALSE;
+    }
     return eglMakeCurrent(display_, EGL_NO_SURFACE, EGL_NO_SURFACE, context_);
 }
 
 EGLBoolean EglManager::EglMakeCurrent(EGLSurface draw, EGLSurface read)
 {
-    CHECK_NOT_INIT(EGL_FALSE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_FALSE;
+    }
     return eglMakeCurrent(display_, draw, read, context_);
 }
 
 EGLBoolean EglManager::EglMakeCurrent(EGLSurface draw, EGLSurface read, EGLContext ctx)
 {
-    CHECK_NOT_INIT(EGL_FALSE);
+    if (!initFlag_) {
+        BLOGE("EglManager is not init.");
+        return EGL_FALSE;
+    }
     return eglMakeCurrent(display_, draw, read, ctx);
 }
-
 } // namespace OHOS
